Read Data from Folder

Reads in all data files of specified types found in the designated folder. Returns a list with one element for each data file. Currently supports CSV, TXT, XLS, and XLSX. This function was built by the following authors: Nicholas Lyon

Usage

read(raw_folder = NULL, data_formats = c("csv", "txt", "xls", "xlsx"))

Arguments

raw_folder

(character) folder / folder path containing data files to read

data_formats

(character) file extensions to identify within the `raw_folder`. Default behavior is to search for all supported file types.

Value

(list) data found in specified folder of specified file format(s)

Examples

if (FALSE) {
# Read in all CSV files in the "raw_data" folder
df_list <- read(raw_folder = "raw_data", data_formats = "csv")
}